# Conflicts:
#	intel_extension_for_transformers/llm/runtime/graph/models/model_utils/model_files.h

+++ intel_extension_for_transformers/neural_chat/tests/nightly/plugins/video/test_sadtalker.py
@@ -22,6 +22,7 @@ import shutil
 import requests
 import torch
 import subprocess
+from intel_extension_for_transformers.neural_chat.utils.common import get_device_type
 
 class TestSadTalker(unittest.TestCase):
     @classmethod
@@ -50,16 +51,18 @@ class TestSadTalker(unittest.TestCase):
 
     @classmethod
     def tearDownClass(self):
-        os.remove(self.output_video_path)
-        os.remove(self.source_image)
-        os.remove(self.driven_audio)
+        os.remove(self.output_video_path) if os.path.exists(self.output_video_path) else None
+        os.remove(self.source_image) if os.path.exists(self.source_image) else None
+        os.remove(self.driven_audio) if os.path.exists(self.driven_audio) else None
         shutil.rmtree(self.checkpoint_dir, ignore_errors=True)
         shutil.rmtree(self.enhancer_dir, ignore_errors=True)
 
+    @unittest.skipIf(get_device_type() != 'cpu', "Only run this test on CPU")
     def test_sadtalker_without_enhancer(self):
         self.sadtalker.convert(source_image=self.source_image, driven_audio=self.driven_audio)
         self.assertTrue(os.path.exists(self.output_video_path))
 
+    @unittest.skipIf(get_device_type() != 'cpu', "Only run this test on CPU")
     def test_sadtalker_with_enhancer(self):
         self.sadtalker.enhancer = 'gfpgan'
         self.sadtalker.convert(source_image=self.source_image, driven_audio=self.driven_audio)
